wperf: Replace all new/delete with smart pointers
Wperf currently uses new/delete to allocate/deallocate heap memory which is
known to be error prone. There are actually several instances of memory
leaks (i.e., new without matching delete) in the wperf implementation
already. This patch reimplements the same with smart pointers and therefore the
memory leaks are fixed at the same time as a result of this change.