wperf-driver: Do not free core_info if it's a NULL pointer
See merge request Linaro/WindowsPerf/windowsperf!62

=== ChangeLog ===

* wperf-driver: Do not free core_info if it's a NULL pointer

Under certain circumstances, `core_info` may not be a valid pointer to
memory. For example, if the call to `HalAllocateHardwareCounters` in
`WindowsPerfDeviceCreate` fails somehow, `core_info` will not be a valid pointer
to memory initialized by `ExAllocatePool2`. Therefore, when the device is
unloaded, we need to make sure not to free `core_info` if it's a NULL pointer.