wperf: Remove backslashes from help message
See merge request Linaro/WindowsPerf/windowsperf!74

=== ChangeLog ===

* wperf: Remove backslashes from help message

The help message currently prints the backslashes as is (e.g., \"imix\") and
they are no longer escapes in the raw string. This patch cleans up these
unwanted characters. While we are here, we also replace double quotes with
single quotes to be consistent.