wperf: fix how we display L3 cache metrics
See merge request Linaro/WindowsPerf/windowsperf!101

=== ChangeLog ===

* wperf: fix how we display L3 cache metrics

This patch relates to https://linaro.atlassian.net/browse/WPERF-135

There is an issue with how we are enumerating events[] and evts[] in "L3
cache metrics:" code (both sections).
We are, in debug mode catching (as we should) an assert telling us
there's a std::vector indexing error.

Debugging shows that events[j - 1] was indexed with j == 0. This is
because `evt_num` value stores value equal to:

    fixed_counter_number + events_specific_counter_number

Index `j` is counting from 0 but zero is `cycles` (fixed) counter. In
this fix we are not indexing all fixed counters (in our case defined
with new macro FIXED_COUNTERS_NO).

Tested on ARM64 machine with with L3 cache events.

* wperf-common: add new macro to describe how many fixed counters we currently support