wperf-driver: add safety net timer for SAMPLE_GET request
When we sent PMU_CTL_SAMPLE_GET it's being set to pending and wait for
sampling buffer to fill.
In original prototype this was done in a way that sampling
PMU_CTL_SAMPLE_GET request was blocked until samples arrive.

This patch introduces very simple static timer `timer_cnt` which is
incrementing in case of SAMPLE_GET being handled. Zeroed when it's
completed.

We will adjust timer value, current 100 to most optimal value.