Fix installation issue when some counters out of numGPC aren't allocatable
See merge request Linaro/WindowsPerf/windowsperf!223

=== ChangeLog ===

* Skip padding tests if assumption of 5 gpc's is not met

* Expose the total gpc_num (not just the free gpc_num) to userspace

* Reuse event_enable/disable_irq by changing them to core_counter_enable/disable_irq_helper

* Convert gpc_num's that needed to be changed to counter_idx_map[gpc_num]

* Fix up functions that needed to go thru counter_idx_map

Also removed core_write_counter_helper. Instead, just get the
actual idx in core_write_counter.

* Make misc changes

* Clean up any memory allocations

* Convert numGPC to numFreeGPC even for hw cfg

* Convert all uses of numGPC to numFreeGPC except when querying the hw cfg

Fixes debug assertion hit when running `wperf.exe list`. wperf checks
if gpc_num >= 5 when defining the builtin metrics. If the numFreeGPC is
less than 5, the list of builtin metrics is empty which triggers a
debug assertion when printing the metrics table.

* Convert all calls to go thru counter idx hashtable

* Create counter idx hashtable to map 0..N to allocatable counter indices

Previously, we were trying to allocate all the counters which might not
be possible. So, instead, we allocate the counters that we find to be
allocatable.