[wperf-lib] Generate wperf-lib as a dynamic library
See merge request Linaro/WindowsPerf/windowsperf!297

=== ChangeLog ===

* [wperf-lib] Generate wperf-lib as a dynamic library

Currently, we produce wperf-lib as a static library (i.e., wperf-lib.a). This
patch changes the project to generate wperf-lib as a dynamic library (i.e.,
wperf-lib.dll).

Note that with this new build configuration, there will still be a wperf-lib.a
generated along with wperf-lib.dll. However, this wperf-lib.a is the "import
library" and should not be confused with the static library we generate before
this change.